backend/requirements.txt
fastapi
uvicorn
docker
python-dotenv
httpx
psycopg2-binary
sqlalchemy
bcrypt
pyjwt
python-multipart
torch
transformers
accelerate
safetensors
